back/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
gunicorn==21.2.0
openpyxl==3.1.2
packaging==23.2
pexpect==4.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
python-docx==1.1.0
openpyxl==3.1.2 
python-pptx==0.6.23
pdfplumber==0.11.0
pdfkit==1.0.0
mammoth==1.7.1
xlsx2html==0.6.1 
docx2pdf==0.1.8
selenium==4.9.1
beautifulsoup4==4.12.2
webdriver_manager==4.0.0
django-filter==24.3
lxml
pytest
requests
pyautogui
pandas
xlrd
xlwt
pyexcel
pyexcel-xlsx
adactivexlsx2html
reportlab





